Add jest tests for app and db
65% code coverage reached. Continue to write tests for both app, db, and server. To test loadData.js, wrap contents in a function to export to a test file.
