+++ middleSectionDetails/details/package.json
@@ -29,6 +29,7 @@
     "helmet": "^3.13.0",
     "mongoose": "^5.2.13",
     "morgan": "^1.9.0",
+    "nodemon": "^1.18.4",
     "webpack": "^4.17.2",
     "webpack-cli": "^3.1.0"
   },
@@ -54,6 +55,8 @@
     "regenerator-runtime": "^0.12.1",
     "sinon": "^6.3.1",
     "style-loader": "^0.23.0",
+    "superagent": "^3.8.3",
+    "supertest": "^3.3.0",
     "webpack-combine-loaders": "^2.0.4",
     "webpack-dev-server": "^3.1.7"
   },
