+++ middleSectionDetails/details/src/db/test/testConfig.js
@@ -1,0 +1,15 @@
+const mongoose = require('mongoose');
+const { Movie, MovieReview } = require('../models');
+
+module.exports = {
+  mongoose,
+  connect: (done) => {
+    return mongoose.connect('mongodb://localhost/fMDB')
+      .then(() => Movie.deleteMany({}))
+      .then(() => MovieReview.deleteMany({}))
+      .then(() => done());
+  },
+  disconnect: (done) => {
+    return mongoose.disconnect(done);
+  }
+}
