+++ middleSectionDetails/details/src/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html>
+  <head>
+    <title>fmdb details</title>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+    <link rel="stylesheet" href="./styles.css" >
+  </head>
+  <body>
+    <div id="root"></div>
+    <script src="./bundle.js"></script>
+  </body>
+</html>
